requirements.txt
fastapi
uvicorn
psycopg2-binary
twilio
openai>=1.0.0
python-multipart
